+++ inc/extras.php
@@ -84,4 +84,31 @@ if ( ! function_exists( 'onepress_get_media_url' ) ) {
     }
 }
 
+/**
+ *  Same hook for the_content
+ * @TODO: do not effect content by plugins
+ *
+ * 8 WP_Embed:run_shortcode
+ * 8 WP_Embed:autoembed
+ * 10 wptexturize
+ * 10 wpautop
+ * 10 shortcode_unautop
+ * 10 prepend_attachment
+ * 10 wp_make_content_images_responsive
+ * 11 capital_P_dangit
+ * 11 do_shortcode
+ * 20 convert_smilies
+ */
+global $wp_embed;
+add_filter( 'onepress_the_content', array( $wp_embed, 'run_shortcode' ), 8 );
+add_filter( 'onepress_the_content', array( $wp_embed, 'autoembed' ), 8 );
+add_filter( 'onepress_the_content', 'wptexturize' );
+add_filter( 'onepress_the_content', 'wpautop' );
+add_filter( 'onepress_the_content', 'shortcode_unautop' );
+add_filter( 'onepress_the_content', 'prepend_attachment' );
+add_filter( 'onepress_the_content', 'wp_make_content_images_responsive' );
+add_filter( 'onepress_the_content', 'capital_P_dangit' );
+add_filter( 'onepress_the_content', 'do_shortcode' );
+add_filter( 'onepress_the_content', 'convert_smilies' );
+