+++ section-parts/section-counter.php
@@ -24,7 +24,7 @@ if ( ! empty ( $boxes ) ) {
                     <?php if ($onepress_counter_subtitle != '') echo '<h5 class="section-subtitle">' . esc_html($onepress_counter_subtitle) . '</h5>'; ?>
                     <?php if ($onepress_counter_title != '') echo '<h2 class="section-title">' . esc_html($onepress_counter_title) . '</h2>'; ?>
                     <?php if ( $desc ) {
-                        echo '<div class="section-desc">' . apply_filters( 'the_content', wp_kses_post( $desc ) ) . '</div>';
+                        echo '<div class="section-desc">' . apply_filters( 'onepress_the_content', wp_kses_post( $desc ) ) . '</div>';
                     } ?>
                 </div>
                 <?php } ?>